Package the schedule validator so clinic repos consume it via pre-commit
Moves the validator, JSON Schema, and role vocabulary into an installable
ici3d_schedule package with a validate-schedule console script and a
.pre-commit-hooks.yaml. Clinic repos can now pin this repo as a pre-commit
hook and validate their schedule data against the shared, versioned schema
without vendoring it (no submodule needed):

  repos:
    - repo: https://github.com/ICI3D/ICI3D.github.io
      rev: <sha>
      hooks:
        - id: validate-schedule

Schema and roles ship as package data resolved relative to the module, so
the same code works pip-installed (distributed hook) and from a repo checkout
(this repo's own hook, CI, transitional composite action). --people-dir now
defaults to _data/team in the working repo, matching the renderer's
site.data.team lookup: instructor records are clinic-owned. tools/
validate_schedule.py is kept as a thin shim for backward compatibility.
