+++ ici3d_schedule/data/schedule-cohort.schema.json
